requirements_complete.txt
# DALRN Complete Requirements - 100% PRD Compliance
# All dependencies for production deployment

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
uvloop==0.19.0
httptools==0.6.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
pyjwt==2.8.0
bcrypt==4.1.1
python-multipart==0.0.6
passlib==1.7.4

# Caching & Performance
redis==5.0.1
aiocache==0.12.2
orjson==3.9.10
hiredis==2.2.3

# Blockchain
web3==6.11.3
eth-account==0.10.0
eth-utils==2.3.1

# Async & Concurrency
aiohttp==3.9.1
asyncio==3.4.3
aiofiles==23.2.1

# Machine Learning
torch>=2.0.0
torch-geometric>=2.3.0
scikit-learn==1.3.2
numpy==1.24.3
networkx==3.2.1

# Homomorphic Encryption
tenseal==0.3.14

# Vector Search
faiss-cpu==1.7.4

# Game Theory
nashpy==0.0.40

# Federated Learning
flwr==1.5.0

# Cryptography
cryptography==41.0.7

# IPFS
ipfshttpclient==0.8.0a2

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
locust==2.17.0

# Utilities
pydantic==2.5.2
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0